Move downloadfile to DAL, general clean up and formatting fixes.
