+++ MCLEndUser/MCLEndUser.BLL.Interfaces/src/main/java/net/toastynetworks/MCLEndUser/BLL/Interfaces/IModpackLogic.java
@@ -2,9 +2,11 @@ package net.toastynetworks.MCLEndUser.BLL.Interfaces;
 
 import net.toastynetworks.MCLEndUser.Domain.Modpack;
 
+import java.io.IOException;
 import java.util.List;
 
 public interface IModpackLogic {
     List<Modpack> GetAllModpacks();
-    void DownloadModpack(String fileUri);
+
+    void downloadFile(String fileURL, String saveDir) throws IOException;
 }
