+++ MCLEndUser/MCLEndUser.BLL/MCLEndUser.BLL.iml
@@ -12,5 +12,14 @@
     <orderEntry type="module" module-name="MCLEndUser.DAL.Interfaces" />
     <orderEntry type="module" module-name="MCLEndUser.Domain" />
     <orderEntry type="module" module-name="MCLEndUser.BLL.Interfaces" />
+    <orderEntry type="module-library">
+      <library name="Maven: net.toastynetworks.MCLSharedUtils:MCLSharedUtils:1.2-SNAPSHOT">
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../MCLSharedUtils/target/MCLSharedUtils-1.2-SNAPSHOT.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
   </component>
 </module>
