+++ MCLEndUser/MCLEndUser.BLL/src/main/java/net/toastynetworks/MCLEnduser/BLL/ModpackLogic.java
@@ -4,6 +4,7 @@ import net.toastynetworks.MCLEndUser.BLL.Interfaces.IModpackLogic;
 import net.toastynetworks.MCLEndUser.DAL.Interfaces.IModpackRepository;
 import net.toastynetworks.MCLEndUser.Domain.Modpack;
 
+import java.io.IOException;
 import java.util.List;
 
 public class ModpackLogic implements IModpackLogic {
@@ -19,7 +20,7 @@ public class ModpackLogic implements IModpackLogic {
     }
 
     @Override
-    public void DownloadModpack(String fileUri) {
-        modpackRepository.DownloadModpack(fileUri);
+    public void downloadFile(String fileURL, String saveDir) throws IOException {
+        modpackRepository.downloadFile(fileURL, saveDir);
     }
 }
