+++ MCLEndUser/MCLEndUser.DAL.Interfaces/src/main/java/net/toastynetworks/MCLEndUser/DAL/Interfaces/IModpackRepository.java
@@ -2,9 +2,11 @@ package net.toastynetworks.MCLEndUser.DAL.Interfaces;
 
 import net.toastynetworks.MCLEndUser.Domain.Modpack;
 
+import java.io.IOException;
 import java.util.List;
 
 public interface IModpackRepository {
     List<Modpack> GetAllModpacks();
-    void DownloadModpack(String fileUri);
+
+    void downloadFile(String fileURL, String saveDir) throws IOException;
 }
