+++ MCLEndUser/MCLEndUser.DAL/src/main/java/net/toastynetworks/MCLEndUser/DAL/Contexts/Interfaces/IModpackContext.java
@@ -2,10 +2,12 @@ package net.toastynetworks.MCLEndUser.DAL.Contexts.Interfaces;
 
 import net.toastynetworks.MCLEndUser.Domain.Modpack;
 
+import java.io.IOException;
 import java.util.List;
 
 public interface IModpackContext {
 
     List<Modpack> GetAllModpacks();
-    void DownloadModpack(String fileUri);
+
+    void downloadFile(String fileURL, String saveDir) throws IOException;
 }
