+++ MCLEndUser/MCLEndUser.DAL/src/main/java/net/toastynetworks/MCLEndUser/DAL/Contexts/ModpackRestApiContext.java
@@ -1,44 +1,26 @@
 package net.toastynetworks.MCLEndUser.DAL.Contexts;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.mashape.unirest.http.HttpResponse;
-import com.mashape.unirest.http.ObjectMapper;
 import com.mashape.unirest.http.Unirest;
-import com.mashape.unirest.request.GetRequest;
 import net.toastynetworks.MCLEndUser.DAL.Contexts.Interfaces.IModpackContext;
 import net.toastynetworks.MCLEndUser.Domain.Modpack;
+import net.toastynetworks.unirest.UnirestObjectMapperUtils;
 
+import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
+import java.net.HttpURLConnection;
+import java.net.URL;
 import java.util.Arrays;
 import java.util.List;
 
 public class ModpackRestApiContext implements IModpackContext {
-    static {
-        Unirest.setObjectMapper(new ObjectMapper() {
-            private com.fasterxml.jackson.databind.ObjectMapper jacksonObjectMapper = new com.fasterxml.jackson.databind.ObjectMapper();
-
-            public <T> T readValue(String value, Class<T> valueType) {
-                try {
-                    return jacksonObjectMapper.readValue(value, valueType);
-                } catch (IOException e) {
-                    throw new RuntimeException(e);
-                }
-            }
-
-            public String writeValue(Object value) {
-                try {
-                    return jacksonObjectMapper.writeValueAsString(value);
-                } catch (JsonProcessingException e) {
-                    throw new RuntimeException(e);
-                }
-            }
-        });
-    }
-
+    UnirestObjectMapperUtils unirestObjectMapperUtils = new UnirestObjectMapperUtils();
 
     @Override
     public List<Modpack> GetAllModpacks() {
-        try{
+        try {
             HttpResponse<Modpack[]> modpackListResponse = Unirest.get("http://localhost:8080/v1/modpack").asObject(Modpack[].class);
             Modpack[] modpackObjectArray = modpackListResponse.getBody();
             return Arrays.asList(modpackObjectArray);
@@ -49,12 +31,65 @@ public class ModpackRestApiContext implements IModpackContext {
         return null;
     }
 
+    /**
+     * Downloads a file from a URL
+     *
+     * @param fileURL HTTP URL of the file to be downloaded
+     * @param saveDir path of the directory to save the file
+     * @throws IOException
+     */
     @Override
-    public void DownloadModpack(String fileUri) {
-        try{
-            Unirest.get(fileUri);
-        } catch (Exception e) {
-            System.out.println(e);
+    public void downloadFile(String fileURL, String saveDir) throws IOException {
+        int BUFFER_SIZE = 4096;
+        URL url = new URL(fileURL);
+        HttpURLConnection httpConn = (HttpURLConnection) url.openConnection();
+        int responseCode = httpConn.getResponseCode();
+
+        // always check HTTP response code first
+        if (responseCode == HttpURLConnection.HTTP_OK) {
+            String fileName = "";
+            String disposition = httpConn.getHeaderField("Content-Disposition");
+            String contentType = httpConn.getContentType();
+            int contentLength = httpConn.getContentLength();
+
+            if (disposition != null) {
+                // extracts file name from header field
+                int index = disposition.indexOf("filename=");
+                if (index > 0) {
+                    fileName = disposition.substring(index + 10,
+                            disposition.length() - 1);
+                }
+            } else {
+                // extracts file name from URL
+                fileName = fileURL.substring(fileURL.lastIndexOf("/") + 1,
+                        fileURL.length());
+            }
+
+            System.out.println("Content-Type = " + contentType);
+            System.out.println("Content-Disposition = " + disposition);
+            System.out.println("Content-Length = " + contentLength);
+            System.out.println("fileName = " + fileName);
+
+            // opens input stream from the HTTP connection
+            InputStream inputStream = httpConn.getInputStream();
+            String saveFilePath = saveDir + File.separator + fileName;
+
+            // opens an output stream to save into file
+            FileOutputStream outputStream = new FileOutputStream(saveFilePath);
+
+            int bytesRead = -1;
+            byte[] buffer = new byte[BUFFER_SIZE];
+            while ((bytesRead = inputStream.read(buffer)) != -1) {
+                outputStream.write(buffer, 0, bytesRead);
+            }
+
+            outputStream.close();
+            inputStream.close();
+
+            System.out.println("File downloaded");
+        } else {
+            System.out.println("No file to download. Server replied HTTP code: " + responseCode);
         }
+        httpConn.disconnect();
     }
 }
