+++ MCLEndUser/MCLEndUser.Domain/src/main/java/net/toastynetworks/MCLEndUser/Domain/Modpack.java
@@ -8,6 +8,14 @@ public class Modpack {
     private String downloadUrl;
 
 
+    public Modpack(String name, String versionType) {
+        this.name = name;
+        this.versionType = versionType;
+    }
+
+    public Modpack() {
+    }
+
     public int getId() {
         return id;
     }
@@ -15,13 +23,6 @@ public class Modpack {
     public void setId(int id) {
         this.id = id;
     }
-
-    public Modpack(String name, String versionType) {
-        this.name = name;
-        this.versionType = versionType;
-    }
-
-    public Modpack() { }
 
     public String getName() {
         return name;
