+++ MCLEndUser/MCLEndUser.Domain/src/main/java/net/toastynetworks/MCLEndUser/Domain/UploadedFile.java
@@ -13,6 +13,10 @@ public class UploadedFile {
         this.size = size;
     }
 
+    public UploadedFile() {
+
+    }
+
     public String getFileName() {
         return fileName;
     }
