+++ MCLEndUser/MCLEndUser.Factory/MCLEndUser.Factory.iml
@@ -26,5 +26,14 @@
     <orderEntry type="library" name="Maven: org.json:json:20160212" level="project" />
     <orderEntry type="module" module-name="MCLEndUser.BLL" />
     <orderEntry type="module" module-name="MCLEndUser.BLL.Interfaces" />
+    <orderEntry type="module-library">
+      <library name="Maven: net.toastynetworks.MCLSharedUtils:MCLSharedUtils:1.2-SNAPSHOT">
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/../../MCLSharedUtils/target/MCLSharedUtils-1.2-SNAPSHOT.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
+      </library>
+    </orderEntry>
   </component>
 </module>
