--- MCLEndUser/MCLEndUser.UI/src/main/java/net/toastynetworks/MCLEndUser/UI/Controller.java
@@ -1,118 +1,0 @@
-package net.toastynetworks.MCLEndUser.UI;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
-import javafx.scene.control.ListView;
-import net.toastynetworks.MCLEndUser.BLL.Interfaces.IModpackLogic;
-import net.toastynetworks.MCLEndUser.Domain.Modpack;
-import net.toastynetworks.MCLEndUser.Factory.ModpackFactory;
-import org.apache.commons.io.FileUtils;
-
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.HttpURLConnection;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.util.List;
-import java.util.ResourceBundle;
-
-public class Controller implements Initializable {
-
-    @FXML
-    private ListView modpackList;
-    private List<Modpack> modpackLists;
-
-    private ObservableList<String> items = FXCollections.observableArrayList();
-    private IModpackLogic modpackLogic = ModpackFactory.CreateLogic();
-
-    private static final int BUFFER_SIZE = 4096;
-
-    /**
-     * Downloads a file from a URL
-     * @param fileURL HTTP URL of the file to be downloaded
-     * @param saveDir path of the directory to save the file
-     * @throws IOException
-     */
-    public static void downloadFile(String fileURL, String saveDir)
-            throws IOException {
-        URL url = new URL(fileURL);
-        HttpURLConnection httpConn = (HttpURLConnection) url.openConnection();
-        int responseCode = httpConn.getResponseCode();
-
-        // always check HTTP response code first
-        if (responseCode == HttpURLConnection.HTTP_OK) {
-            String fileName = "";
-            String disposition = httpConn.getHeaderField("Content-Disposition");
-            String contentType = httpConn.getContentType();
-            int contentLength = httpConn.getContentLength();
-
-            if (disposition != null) {
-                // extracts file name from header field
-                int index = disposition.indexOf("filename=");
-                if (index > 0) {
-                    fileName = disposition.substring(index + 10,
-                            disposition.length() - 1);
-                }
-            } else {
-                // extracts file name from URL
-                fileName = fileURL.substring(fileURL.lastIndexOf("/") + 1,
-                        fileURL.length());
-            }
-
-            System.out.println("Content-Type = " + contentType);
-            System.out.println("Content-Disposition = " + disposition);
-            System.out.println("Content-Length = " + contentLength);
-            System.out.println("fileName = " + fileName);
-
-            // opens input stream from the HTTP connection
-            InputStream inputStream = httpConn.getInputStream();
-            String saveFilePath = saveDir + File.separator + fileName;
-
-            // opens an output stream to save into file
-            FileOutputStream outputStream = new FileOutputStream(saveFilePath);
-
-            int bytesRead = -1;
-            byte[] buffer = new byte[BUFFER_SIZE];
-            while ((bytesRead = inputStream.read(buffer)) != -1) {
-                outputStream.write(buffer, 0, bytesRead);
-            }
-
-            outputStream.close();
-            inputStream.close();
-
-            System.out.println("File downloaded");
-        } else {
-            System.out.println("No file to download. Server replied HTTP code: " + responseCode);
-        }
-        httpConn.disconnect();
-    }
-
-    public void playButtonClicked(){
-        try {
-            Modpack modpack = modpackLists.stream().filter(x -> x.getName() == modpackList.getSelectionModel().getSelectedItem()).findFirst().get();
-            System.out.println(modpack.getName());
-            //TODO: Get download URL here from the object and then make a rest call to retrieve the file.
-//        modpackLogic.DownloadModpack(modpack.getDownloadUrl());
-            //Temporary
-            downloadFile(modpack.getDownloadUrl(),"C:\\Users\\user\\AppData\\Roaming\\.MCLauncher\\test");
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    @Override
-    public void initialize(URL location, ResourceBundle resources) {
-        modpackLists = modpackLogic.GetAllModpacks();
-        modpackList.setItems(items);
-        for (Modpack modpack :
-                modpackLists) {
-            items.add(modpack.getName());
-        }
-
-    }
-}
